DC-431 Fix PropertyName for location in blueprint.xml
