+++ minion/minion-flows/flows-parser/src/main/java/org/opennms/horizon/minion/flows/parser/FlowSinkModule.java
@@ -29,6 +29,7 @@
 package org.opennms.horizon.minion.flows.parser;
 
 import com.google.protobuf.InvalidProtocolBufferException;
+import org.opennms.horizon.minion.flows.parser.factory.FlowsIdentity;
 import org.opennms.cloud.grpc.minion.Identity;
 import org.opennms.horizon.grpc.telemetry.contract.TelemetryMessage;
 import org.opennms.horizon.grpc.telemetry.contract.TelemetryMessageLog;
@@ -39,9 +40,9 @@ import org.opennms.horizon.shared.ipc.sink.api.SinkModule;
 
 public class FlowSinkModule implements SinkModule<TelemetryMessage, TelemetryMessageLog> {
 
-    private final org.opennms.horizon.minion.flows.parser.factory.Identity identity;
+    private final FlowsIdentity identity;
 
-    public FlowSinkModule(org.opennms.horizon.minion.flows.parser.factory.Identity identity) {
+    public FlowSinkModule(FlowsIdentity identity) {
         this.identity = identity;
     }
