+++ minion/minion-flows/flows-parser/src/main/java/org/opennms/horizon/minion/flows/parser/IpfixUdpParser.java
@@ -38,7 +38,7 @@ import org.opennms.horizon.grpc.telemetry.contract.TelemetryMessage;
 import org.opennms.horizon.minion.flows.listeners.Dispatchable;
 import org.opennms.horizon.minion.flows.listeners.UdpParser;
 import org.opennms.horizon.minion.flows.parser.factory.DnsResolver;
-import org.opennms.horizon.minion.flows.parser.factory.Identity;
+import org.opennms.horizon.minion.flows.parser.factory.FlowsIdentity;
 import org.opennms.horizon.minion.flows.parser.ie.RecordProvider;
 import org.opennms.horizon.minion.flows.parser.ipfix.proto.Header;
 import org.opennms.horizon.minion.flows.parser.ipfix.proto.Packet;
@@ -60,7 +60,7 @@ public class IpfixUdpParser extends UdpParserBase implements UdpParser, Dispatch
 
     public IpfixUdpParser(final String name,
                           final AsyncDispatcher<TelemetryMessage> dispatcher,
-                          final Identity identity,
+                          final FlowsIdentity identity,
                           final DnsResolver dnsResolver,
                           final MetricRegistry metricRegistry) {
         super(Protocol.IPFIX, name, dispatcher, identity, dnsResolver, metricRegistry);
