+++ minion/minion-flows/flows-parser/src/main/java/org/opennms/horizon/minion/flows/parser/UdpParserBase.java
@@ -36,7 +36,7 @@ import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
 
 import org.opennms.horizon.minion.flows.parser.factory.DnsResolver;
-import org.opennms.horizon.minion.flows.parser.factory.Identity;
+import org.opennms.horizon.minion.flows.parser.factory.FlowsIdentity;
 import org.opennms.horizon.minion.flows.parser.ie.RecordProvider;
 import org.opennms.horizon.minion.flows.parser.session.Session;
 import org.opennms.horizon.minion.flows.parser.session.UdpSessionManager;
@@ -50,7 +50,6 @@ import com.codahale.metrics.MetricRegistry;
 
 import io.netty.buffer.ByteBuf;
 import org.opennms.horizon.minion.flows.listeners.UdpParser;
-import org.opennms.horizon.minion.flows.listeners.factory.UdpListenerMessage;
 
 public abstract class UdpParserBase extends ParserBase implements UdpParser {
     public final static long HOUSEKEEPING_INTERVAL = 60000;
@@ -66,7 +65,7 @@ public abstract class UdpParserBase extends ParserBase implements UdpParser {
     public UdpParserBase(final Protocol protocol,
                          final String name,
                          final AsyncDispatcher<TelemetryMessage> dispatcher,
-                         final Identity identity,
+                         final FlowsIdentity identity,
                          final DnsResolver dnsResolver,
                          final MetricRegistry metricRegistry) {
         super(protocol, name, dispatcher, identity, dnsResolver, metricRegistry);
