+++ minion/minion-flows/flows-parser/src/test/java/org/opennms/horizon/minion/flows/parser/ClockSkewTest.java
@@ -37,7 +37,7 @@ import org.junit.Before;
 import org.junit.Ignore;
 import org.opennms.horizon.grpc.telemetry.contract.TelemetryMessage;
 import org.opennms.horizon.minion.flows.parser.factory.DnsResolver;
-import org.opennms.horizon.minion.flows.parser.factory.Identity;
+import org.opennms.horizon.minion.flows.parser.factory.FlowsIdentity;
 import org.opennms.horizon.minion.flows.parser.flowmessage.FlowMessage;
 import org.opennms.horizon.minion.flows.parser.transport.MessageBuilder;
 import org.opennms.horizon.shared.ipc.sink.api.AsyncDispatcher;
@@ -47,7 +47,7 @@ import com.codahale.metrics.MetricRegistry;
 public class ClockSkewTest {
     private int eventCount = 0;
 
-    private final Identity identity = new Identity() {
+    private final FlowsIdentity identity = new FlowsIdentity() {
         @Override
         public String getId() {
             return "myId";
@@ -142,7 +142,7 @@ public class ClockSkewTest {
 
     private static class ParserBaseExt extends ParserBase {
 
-        public ParserBaseExt(Protocol protocol, String name, AsyncDispatcher<TelemetryMessage> dispatcher, Identity identity, DnsResolver dnsResolver, MetricRegistry metricRegistry) {
+        public ParserBaseExt(Protocol protocol, String name, AsyncDispatcher<TelemetryMessage> dispatcher, FlowsIdentity identity, DnsResolver dnsResolver, MetricRegistry metricRegistry) {
             super(protocol, name, dispatcher, identity, dnsResolver, metricRegistry);
         }
